[Sparse] Move `SparseTensorUtils.*` to `native/`
