+++ aten/src/ATen/SparseCsrTensorImpl.cpp
@@ -3,7 +3,6 @@
 #include <ATen/SparseCsrTensorImpl.h>
 #include <ATen/SparseCsrTensorUtils.h>
 #include <ATen/SparseTensorImpl.h>
-#include <ATen/SparseTensorUtils.h>
 #include <ATen/core/LegacyTypeDispatch.h>
 #include <ATen/native/Resize.h>
