+++ aten/src/ATen/native/sparse/SoftMax.cpp
@@ -4,8 +4,8 @@
 #include <ATen/Dispatch.h>
 #include <ATen/NamedTensorUtils.h>
 #include <ATen/native/sparse/ParamUtils.h>
+#include <ATen/native/SparseTensorUtils.h>
 #include <ATen/Parallel.h>
-#include <ATen/SparseTensorUtils.h>
 #include <c10/util/accumulate.h>
 #include <c10/util/irange.h>
