+++ aten/src/ATen/native/sparse/SparseUnaryOps.cpp
@@ -1,5 +1,5 @@
 // #define TORCH_ASSERT_ONLY_METHOD_OPERATORS
-#include <ATen/SparseTensorUtils.h>
+#include <ATen/native/SparseTensorUtils.h>
 #include <ATen/core/Tensor.h>
 
 #ifndef AT_PER_OPERATOR_HEADERS
