+++ aten/src/ATen/native/sparse/cuda/SparseCsrTensorMath.cu
@@ -5,10 +5,10 @@
 #include <ATen/InitialTensorOptions.h>
 #include <ATen/SparseCsrTensorImpl.h>
 #include <ATen/SparseCsrTensorUtils.h>
-#include <ATen/SparseTensorUtils.h>
 #include <ATen/WrapDimUtilsMulti.h>
 #include <ATen/native/BinaryOps.h>
 #include <ATen/native/Resize.h>
+#include <ATen/native/SparseTensorUtils.h>
 #include <algorithm>
 
 #ifndef AT_PER_OPERATOR_HEADERS
