+++ build_variables.bzl
@@ -989,7 +989,6 @@ aten_cpu_source_non_codegen_list = [
     "aten/src/ATen/ScalarOps.cpp",
     "aten/src/ATen/SparseTensorImpl.cpp",
     "aten/src/ATen/SparseCsrTensorImpl.cpp",
-    "aten/src/ATen/SparseTensorUtils.cpp",
     "aten/src/ATen/TensorGeometry.cpp",
     "aten/src/ATen/TensorIndexing.cpp",
     "aten/src/ATen/TensorMeta.cpp",
@@ -1309,6 +1308,7 @@ aten_native_source_non_codegen_list = [
     "aten/src/ATen/native/SobolEngineOpsUtils.cpp",
     "aten/src/ATen/native/SoftMax.cpp",
     "aten/src/ATen/native/Sorting.cpp",
+    "aten/src/ATen/native/SparseTensorUtils.cpp",
     "aten/src/ATen/native/SpectralOps.cpp",
     "aten/src/ATen/native/SummaryOps.cpp",
     "aten/src/ATen/native/TensorAdvancedIndexing.cpp",
