+++ torch/csrc/distributed/c10d/ProcessGroupGloo.cpp
@@ -34,7 +34,7 @@
 #include <gloo/reduce.h>
 #include <gloo/scatter.h>
 
-#include <ATen/SparseTensorUtils.h>
+#include <ATen/native/SparseTensorUtils.h>
 #include <ATen/ThreadLocalState.h>
 
 #include <c10/util/StringUtil.h>
